# hq/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zako3 HQ API",
    "description": "RESTful API for Zako3 platform - manage users, communities (taps), and bot settings",
    "version": "1.0.0",
    "contact": {
      "name": "Zako3 Team"
    },
    "license": {
      "name": "ISC"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication provider"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/taps/{tapId}/members": {
      "get": {
        "summary": "Get tap members",
        "tags": [
          "taps"
        ],
        "description": "Get list of tap members",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "tapId",
            "required": true,
            "description": "Tap ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tapId": {
                            "type": "string",
                            "description": "Tap ID"
                          },
                          "userId": {
                            "type": "string",
                            "description": "User ID"
                          },
                          "role": {
                            "type": "string",
                            "description": "Member role"
                          },
                          "joinedAt": {
                            "type": "string",
                            "description": "Join timestamp"
                          }
                        },
                        "required": [
                          "tapId",
                          "userId",
                          "role",
                          "joinedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.zako3.com",
      "description": "Production server"
    }
  ],
...
}